fix: authOptions build 오류 수정
