+++ src/app/api/bookmark/route.ts
@@ -1,11 +1,11 @@
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../auth/[...nextauth]/route';
 import { NextRequest, NextResponse } from 'next/server';
 import {
   addBookmark,
   deleteBookmark,
   getBookmark,
 } from '@/service/query/member';
+import { authOptions } from '@/utils/authOptions';
 
 export async function GET() {
   const session = await getServerSession(authOptions);
