+++ src/app/api/elite/reset/route.ts
@@ -1,7 +1,7 @@
 import { getServerSession } from 'next-auth';
 import { NextResponse } from 'next/server';
 import { memberEliteCollectionsReset } from '@/service/query/member';
-import { authOptions } from '../../auth/[...nextauth]/route';
+import { authOptions } from '@/utils/authOptions';
 
 export async function PUT() {
   const session = await getServerSession(authOptions);
