+++ src/app/api/elite/route.ts
@@ -1,11 +1,11 @@
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../auth/[...nextauth]/route';
 import { NextRequest, NextResponse } from 'next/server';
 import {
   getEliteCollections,
   updateEliteCollections,
 } from '@/service/query/member';
 import { EliteCollections } from '@/model/monster';
+import { authOptions } from '@/utils/authOptions';
 
 type EliteList = {
   eliteList: EliteCollections[];
